refactor /network into groups+subgroups
