--- app/page/network.mcss
@@ -1,40 +1,0 @@
-NetworkPage {
-  width: 100%
-
-  display: grid
-  justify-content: center
-
-  div.container {
-    section {
-      margin-bottom: 2rem
-      h2 {
-        font-size: .8rem
-        text-transform: uppercase
-        letter-spacing: 2px
-
-        i {
-          color: #666
-          margin-left: .5rem
-        }
-      }
-
-      div.invite {
-        margin-top: 1rem
-
-        display: grid
-        grid-template-columns: 55rem 8rem 2rem
-        grid-gap: 1rem
-
-        input {
-          padding-left: .5rem
-        }
-
-        i.fa {
-          color: fuchsia
-          justify-self: center
-          align-self: center
-        }
-      }
-    }
-  }
-}
